small adjustments and move of fakeEvents to fakeData folder
